Add observability into cell code when debugging
- Add cell code to Python's linecache when it is compiled, under a temp
  file
- Refactor cell compilation into its own module
